Altered line 312 to include (byte) typecast on CMD_ADC_READ since it was throwing errors when compiling on a Mac install. This shouldn't change operation in any way.

+++ MS5803_01.cpp
@@ -309,7 +309,7 @@ unsigned long MS_5803::MS_5803_ADC(char commandADC) {
     }
     // Now send the read command to the MS5803 
     Wire.beginTransmission(MS5803_I2C_ADDRESS);
-    Wire.write(CMD_ADC_READ);
+    Wire.write((byte)CMD_ADC_READ);
     Wire.endTransmission();
     // Then request the results. This should be a 24-bit result (3 bytes)
     Wire.requestFrom(MS5803_I2C_ADDRESS, 3);
